feat: pause on dialog messages
